profile, profile edit, ahopping cart, shopping cart edit added

+++ src/main/java/com/novo/personalproject/dao/ShoppingCartRepository.java
@@ -3,5 +3,5 @@ package com.novo.personalproject.dao;
 import com.novo.personalproject.model.entity.ShoppingCart;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ShoppingCartRepository extends JpaRepository<ShoppingCart, Integer> {
+public interface ShoppingCartRepository extends JpaRepository<ShoppingCart, Long> {
 }
